+++ src/main/java/org/bioshock/engine/renderers/SwatterRenderer.java
@@ -12,8 +12,8 @@ public final class SwatterRenderer implements Renderer {
         GraphicsContext gc,
         E swatter
     ) {
-        int x = swatter.getX();
-        int y = swatter.getY();
+        double x = swatter.getX();
+        double y = swatter.getY();
         double width = swatter.getWidth();
         double height = swatter.getHeight();
